Minor OpenBSD fixes.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@3102 42547f41-83fc-0310-9aef-9668e019ab9e

+++ motor/mshow.c
@@ -1030,7 +1030,8 @@ motor_show_version( void )
 		list_head->mx_version_time );
 
 
-	fprintf( output, "\nCurrent Posix time: %lu seconds\n", time( NULL ) );
+	fprintf( output, "\nCurrent Posix time: %lu seconds\n",
+					(unsigned long) time( NULL ) );
 
 	fprintf( output, "\n" );
